app/requirements.txt
crawl4ai
cv2-opencv-python-headless
exiftool
fastapi
httpx
pillow
pytesseract
python-multipart
uvicorn[standard]